requirements.txt
matplotlib
numpy
opencv-python-headless
opencv-python
pillow
requests
watchdog
psycopg2-binary